% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT.R
\name{CT}
\alias{CT}
\title{Cross table / Chi-square test}
\usage{
CT(DV, IDV, freqCorrect = 5)
}
\arguments{
\item{DV}{The dependent variable}

\item{IDV}{The independent variable}

\item{freqCorrect}{Minimum observations per cell for Fisher correction - default is 5}
}
\value{
Printed model with and a returned list with the following components:

Effect_size: Cramér's V effect size (if the model is significant)

Figure
}
\description{
This function conducts a Chi-square test with a fisher correction (if needed).
Based on gmodels::CrossTable()
}
\examples{
CT(theData$Gender, theData$Condition)

}
