% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GF.R
\name{GF}
\alias{GF}
\title{Chi-square test for Goodness of Fit}
\usage{
GF(DV, Probs)
}
\arguments{
\item{DV}{The variable to be tested}

\item{Probs}{Vector of probabilities}
}
\value{
Printed model with and a returned list with the following components:

Effect_size: Cohen's W effect size (if the model is significant)

Figure
}
\description{
This function conducts a Chi-square test with a fisher correction (if needed).
Based on gmodels::CrossTable()
}
\examples{
CT(rowFactor = simulateData$Gender)
}
