% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoF.R
\name{GoF}
\alias{GoF}
\title{Chi-square for Goodness of Fit test}
\usage{
GoF(DV, Probs)
}
\arguments{
\item{DV}{The vDependent variable}

\item{Probs}{The probabilities}
}
\value{
Printed model with and a returned list with the following components:

Effect_size: Cohen's W effect size (if the model is significant)

Figure
}
\description{
This function conducts a Chi-square test
Based on gmodels::CrossTable()
}
