% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indttest.R
\name{indttest}
\alias{indttest}
\title{Independent samples t-test}
\usage{
indttest(DV, IDV)
}
\arguments{
\item{DV}{A vector with the independent variable}

\item{IDV}{A vector with the dependent variable}
}
\value{
A list with the descriptive statistics, the model, effect size (if the model is significant) and a figure
}
\description{
The function uses the 'stats' package to conduct the independent sample t-test or the Welch test
The correction for the Welch test is automatically using the Leven test
Based on stats::t.test()
}
\examples{
indttest(theData$Score, theData$Gender)


}
