% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqrOutliers.R
\name{iqrOutliers}
\alias{iqrOutliers}
\title{IQR Outliers}
\usage{
iqrOutliers(data, lowerBound = 2.5, upperBound = 2.5, replace = FALSE)
}
\arguments{
\item{data}{Input vector}

\item{lowerBound}{Lower bound (default value is 2.5)}

\item{upperBound}{Upper bound (default value is 2.5)}

\item{replace}{If replace is TRUE (default is FALSE) than the function also use the replaceWith function with the median default values}
}
\value{
A vector with outliers as NA values
}
\description{
This function takes a vector, and replace observations that are smaller or bigger than the lower/upper bounds with NA.
}
\examples{
iqrOutliers(data = theData$Age)
}
