% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logReg.R
\name{logReg}
\alias{logReg}
\title{Logistic regression}
\usage{
logReg(DV, Predictors, Classification = 0.5)
}
\arguments{
\item{DV}{The dependent variable - a numeric 0-1 vector}

\item{Predictors}{The independent variables}

\item{Classification}{The classification boudery for 1 values (defalut is 0.5)}
}
\value{
The model, odds ratios and confidence intervals
}
\description{
The function conducts logistic regression together with the preliminary tests: NagelKerke and Hosmer tests.
Based on stats::glm(), rcompanion::nagelkerke(), ResourceSelection::hoslem.test
}
\examples{

}
