% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_CI.R
\name{mean_CI}
\alias{mean_CI}
\title{Calculating Mean CI using Student t distribution}
\usage{
mean_CI(data, alpha_threshold = 0.05)
}
\arguments{
\item{alpha_threshold}{}
}
\value{
Return a list with the mean and high and low CI
}
\description{
Calculating Mean CI using Student t distribution
}
