% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiReg.R
\name{multiReg}
\alias{multiReg}
\title{Multiple or single linear regression}
\usage{
multiReg(DV, Predictors, Correction = "HC2")
}
\arguments{
\item{DV}{The dependent variables}

\item{Predictors}{The independent variables (order for string factors will be by the ABC)}

\item{Correction}{The type of correction for heteroscedasticity (default is HC2, for more information see estimatr::lm_robust se_type)}
}
\value{
The model with unstandardized beta coefficients, standardized beta coefficients, VIF and type of correction
}
\description{
The function conduct multiple linear regression.
The type of correction for heteroscedasticity (no correction v.s HC2 or other is done automatically)
If multicolliniarity is exists, the model will run, but the VIF values will be printed with a recommendation to remove variables.
Based on stats::lm(), estimatr::lm_robust(), faraway::vif(), lmtest::bptest()
}
\examples{

}
