% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneSampleTest.R
\name{oneSampleTest}
\alias{oneSampleTest}
\title{One sample t.test}
\usage{
oneSampleTest(DV, MU = 0, Parametric = TRUE)
}
\arguments{
\item{DV}{TA vector with the dependent variable}

\item{MU}{The parameter for comparison}

\item{Parametric}{If FALSE the test is Mann-Whitney test}
}
\value{
A list with the following components:

Descriptive_statistics: Descriptive statistics with the Mean, standard deviation, Median and N

Model_summary: Model's summary

Effect_size: Cohen's d/rank_biserial effect size (if the model is significant and parametric)

Figure
}
\description{
This function conducts a one sample t.test or one sample Mann-Whitney.
}
\examples{
oneSampleTest(DV = simulateData$Score, MU = 5)
}
