% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneWayAnova.R
\name{oneWayAnova}
\alias{oneWayAnova}
\title{One Way Anova}
\usage{
oneWayAnova(DV, IDV, Correction = "BH", Parametric = TRUE)
}
\arguments{
\item{DV}{A vector with the independent variable}

\item{IDV}{A vector with the dependent variable}

\item{Correction}{The type of correction for post hoc (default is Benjamini, Y., and Hochberg, Y. (1995)) for more details go to the function pairwise.t.test}

\item{Parametric}{If FALSE the test is Kruskal-Wallis test with Mann-Whitney test for post hoc pairwise comparisons}
}
\value{
A list with (1) Descriptive statistics, (2) The model, (3) Effect size - if the model is significant, (4) t-test pairwise comparisons with correction - if significant, (5) Figure
}
\description{
This function conducts a tyoe 3 one way ANOVA or with hc2 correction type (if the variances are unequal). If the model is significant, an effect size and pairwise comparisons  also produced
Based on stats::aov(), car::Anova()
}
\examples{
oneWayAnova(theData$Score, theData$Condition)

}
