% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageData.R
\name{packageData}
\alias{packageData}
\title{Package's data calculation
Calculating  observations' ID, Age, Gender, measuring time, Condition in A/B testing, score at the game and time at the game.}
\usage{
packageData(data = simulateData)
}
\value{
A data frame
}
\description{
Package's data calculation
Calculating  observations' ID, Age, Gender, measuring time, Condition in A/B testing, score at the game and time at the game.
}
\examples{
Data <- packageData()
}
