% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedT.R
\name{pairedT}
\alias{pairedT}
\title{Paired samples t-test}
\usage{
pairedT(DV, IDV, ID, Parametric = TRUE)
}
\arguments{
\item{DV}{A vector with the independent variable}

\item{IDV}{A vector with the dependent variable}

\item{ID}{The identity of the participants}

\item{Parametric}{If FALSE the test is Wilcoxon Sum Rank test}
}
\value{
A list with the following components:

Descriptive_statistics: Descriptive statistics with the Mean, standard deviation, Median and N

Model_summary: Model's summary

Effect_size: Cohen's d effect size (if the model is significant and parametric)

Figure
}
\description{
The function conducts paired sample t-test. The data set should be with exact n observations per ID, otherwise the figure's and descriptive statistics' results will be inaccurate.
Based on stats::t.test()
}
\examples{
pairedT(theData$Score, theData$Gender, theData$ID)

}
