% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porportion_CI.R, R/prop_CI.R
\name{porportion_CI}
\alias{porportion_CI}
\title{'do proportion CI}
\usage{
porportion_CI(n, p)

porportion_CI(n, p)
}
\arguments{
\item{p}{}
}
\description{
'do proportion CI

Calculate proportion CI
}
