% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHoc.R
\name{postHoc}
\alias{postHoc}
\title{Post-hoc pairwise t-test comparisons}
\usage{
postHoc(
  DV,
  IDV,
  ID = NULL,
  Correction = "BH",
  Paired = FALSE,
  Parametric = TRUE
)
}
\arguments{
\item{DV}{The dependent variable}

\item{IDV}{The independent variable}

\item{ID}{The identity of the observations (mandatory for paired tests, the default is NULL)}

\item{Correction}{The type of correction (default is Benjamini, Y., and Hochberg, Y. (1995)) for more details go to the function pairwise.t.test}

\item{Paired}{Should be TRUE if the test is for paired comparisons (default is FALSE)}

\item{Parametric}{If TRUE the tests are t.tests, if FALSE the tests are Mann-Whitney/Wilcoxon Sum Rank tests}
}
\value{
The model summary
}
\description{
This function conduct pairwise t-test comparisons.
Based on stats::pairwise.t.test()
}
\examples{
postHoc(theData$Score, theData$Condition)

}
