% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceWith.R
\name{replaceWith}
\alias{replaceWith}
\title{Replace with}
\usage{
replaceWith(data, type = "Mean", Ratio = 0.8)
}
\arguments{
\item{data}{A vector}

\item{type}{Type of statistic (default in the mean)}

\item{Ratio}{Minimum ratio for replacement (default in 80\%)}
}
\value{
New vector with no missing values
}
\description{
This function takes a vector and if the missing values ratio are greater than a given value (0.8 default), its replaced with either the Mean (default), Median or mode
}
\examples{
replaceWith(data = simulateData$gameTime)

}
