% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmAnova.R
\name{rmAnova}
\alias{rmAnova}
\title{Repeated measures One Way Anova}
\usage{
rmAnova(DV, IDV, ID, Correction = "BH")
}
\arguments{
\item{DV}{A vector with the independent variable}

\item{IDV}{A vector with the dependent variable}

\item{ID}{The identity of the observations}

\item{Correction}{he type of correction for post hoc (default is Benjamini, Y., and Hochberg, Y. (1995)) for more details go to the function pairwise.t.test}
}
\value{
A list with (1) Descriptive statistics, (2) The model, (3) Effect size - if the model is significant, (4) t-test pairwise comparisons with correction - if significant, (5) Figure
}
\description{
This function conducts repeated measures ANOVA. If the model is significant, an effect size and pairwise comparisons also be produced. The data set should be with exact n observations per ID, otherwise the figure's and descriptive statistics' results will be inaccurate
Based on stats::aov()
}
\examples{
rmAnova(theData$Score, theData$Condition, theData$ID)

}
