% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdOutliers.R
\name{sdOutliers}
\alias{sdOutliers}
\title{SD (standard deviation) outliers}
\usage{
sdOutliers(data, lowerBound = 3, upperBound = 3, replace = FALSE)
}
\arguments{
\item{data}{A numeric vector}

\item{lowerBound}{Lower bound standard deviation (default value is 3)}

\item{upperBound}{Upper bound standard deviation (default value is 3)}

\item{replace}{If replace is TRUE (default is FALSE) than the function also use the replaceWith function with the default values}
}
\value{
A vector with outliers as NA values
}
\description{
This function takes a vector, and replace observations that are smaller or bigger than the Mean + the lower/upper bounds * standard deviation with NA.
}
\examples{
sdOutliers(data = simulateData$gameTime)

}
