% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulateData}
\alias{simulateData}
\title{Data set that contains the information as data frame}
\format{
A data frame with 600 observations and 7 variables:
\describe{
\item{ID}
\item{Age}
\item{Gender}
\item{measureTime}
\item{Condition}
\item{gameTime}
\item{Score}
}
}
\source{
{Created in house.}
}
\usage{
simulateData
}
\description{
The data frame contains 600 rows with 7 columns.
}
\examples{
data(simulateData)



}
\keyword{datasets}
