% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theData}
\alias{theData}
\title{Data that contains the  information as data frame}
\format{
A data frame with 30 rows and 5 variables:
\describe{
\item{ID}
\item{Age}
\item{Gender}
\item{Time}
\item{gameTime}
\item{Score}
}
}
\source{
{Created in house.}
}
\usage{
theData
}
\description{
The data frame contain 200 observations with 5 columns.
}
\examples{
data(theData)



}
\keyword{datasets}
