% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoWay.R
\name{twoWay}
\alias{twoWay}
\title{Two Way Anova}
\usage{
twoWay(DV, IDV1, IDV2, Correction = "BH")
}
\arguments{
\item{DV}{A vector with the dependent variable}

\item{IDV1}{A vector with the first independent variable}

\item{IDV2}{A vector with the second independent variable}

\item{Correction}{The type of correction for post hoc (default is Benjamini, Y., and Hochberg, Y. (1995)) for more details go to the function pairwise.t.test}
}
\value{
A list with the following components:

Descriptive_statistics: Descriptive statistics with the Mean, standard deviation, Median and N

Model_summary: Model's summary (with or without correction for variance)

Effect_size: eta square effect size (if the model is significant and parametric)

Post_hoc_IDV1: Pairwise comparisons for the first independent variable (if significant)

Post_hoc_IDV2: Pairwise comparisons for the second independent variable (if significant)

Post_hoc_Interaction: Pairwise comparisons for the interaction (if significant)

Variance_Correction: If TRUE a correction for the variance was conducted (the test type is 'Analysis of Deviance')

Figure
}
\description{
This function conducts a type 3 two way ANOVA or with hc2 correction type (if the variances are unequal). If the model is significant, an effect size and pairwise comparisons also produced
Based on stats::aov(), car::Anova()
}
\examples{
twoWay(theData$Score, theData$Condition, theData$Gender)
}
