% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT.R
\name{CT}
\alias{CT}
\title{Cross table / Chi-square for Independence test}
\usage{
CT(rowFactor, colFactor, freqCorrect = 5)
}
\arguments{
\item{colFactor}{The variable to be locaated in the columns}

\item{freqCorrect}{Minimum observations per cell for Fisher correction - default is 5}

\item{firstFactor}{The variable to be locaated in the rows}
}
\value{
Printed model with and a returned list with the following components:

Effect_size: Phi or Cramér's V effect size (if the model is significant)

Figure
}
\description{
This function conducts a Chi-square test with a fisher correction (if needed).
Based on gmodels::CrossTable()
}
\examples{
CT(rowFactor = simulateData$Gender, colFactor = simulateData$Condition)

}
