% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indttest.R
\name{indttest}
\alias{indttest}
\title{Independent samples t-test}
\usage{
indttest(DV, IDV, Parametric = TRUE)
}
\arguments{
\item{DV}{A vector with the dependent variable}

\item{IDV}{A vector with the independent variable}

\item{Parametric}{If FALSE the test is Mann-Whitney test}
}
\value{
A list with the following components:

Descriptive_statistics: Descriptive statistics with the Mean, standard deviation, Median and N

Model_summary: Model's summary (with or without correction for variance)

Effect_size: Cohen's d effect size (if the model is significant and parametric)

Variance_Correction: If TRUE a correction for the variance was conducted (the test type is 'Welch's t.test')

Figure
}
\description{
The function uses the 'stats' package to conduct the independent sample t-test or the Welch test
The correction for the Welch test is automatically using the Leven test
Based on stats::t.test()
}
\examples{
indttest(DV = simulateData$Score, IDV = simulateData$Gender)


}
