% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logReg.R
\name{logReg}
\alias{logReg}
\title{Logistic regression}
\usage{
logReg(DV, Predictors, Classification = 0.5)
}
\arguments{
\item{DV}{The dependent variable - a numeric 0-1 vector. IF the DV is a character type, the function will give the values o and 1 using alphabetic order.}

\item{Predictors}{The independent variables}

\item{Classification}{The classification boundary for 1 values (default is 0.5)}
}
\value{
A list with the following components:

Model_Summary: The model's summary

Odds_Ratio: Odds ratios and 95\% confidence intervals
}
\description{
The function conducts logistic regression together with the preliminary tests: NagelKerke and Hosmer-Lemeshow Goodness of Fit tests.
Based on stats::glm(), rcompanion::nagelkerke(), ResourceSelection::hoslem.test
}
\examples{
logReg(DV = simulateData$Gender, Predictors = simulateData[, c('Age', 'Score')])

}
