% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_CI.R
\name{mean_CI}
\alias{mean_CI}
\title{Calculating Mean CI using Student t distribution}
\usage{
mean_CI(data, conidance_interval = 0.95)
}
\arguments{
\item{data}{input numeric vector}

\item{conidance_interval}{default is 95\%}
}
\value{
Return a list with the mean and high and low CI
}
\description{
Calculating Mean CI using Student t distribution
}
