% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneWayAnova.R
\name{oneWayAnova}
\alias{oneWayAnova}
\title{One Way Anova}
\usage{
oneWayAnova(DV, IDV, Correction = "BH", Parametric = TRUE)
}
\arguments{
\item{DV}{A vector with the dependent variable}

\item{IDV}{A vector with the independent variable}

\item{Correction}{The type of correction for post hoc (default is Benjamini, Y., and Hochberg, Y. (1995)) for more details go to the function pairwise.t.test}

\item{Parametric}{If FALSE the test is Kruskal-Wallis test with Mann-Whitney test for post hoc pairwise comparisons}
}
\value{
A list with the following components:

Descriptive_statistics: Descriptive statistics with the Mean, standard deviation, Median and N

Model_summary: Model's summary (with or without correction for variance)

Effect_size: eta square effect size (if the model is significant and parametric)

Post_hoc: Pairwise comparisons (if the model is significant)

Figure
}
\description{
This function conducts a type 3 one way ANOVA or with hc2 correction type (if the variances are unequal). If the model is significant, an effect size and pairwise comparisons  also produced
Based on stats::aov(), car::Anova()
}
\examples{
oneWayAnova(theData$Score, theData$Condition)

}
