% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageData.R
\name{packageData}
\alias{packageData}
\title{Package's data calculation
Calculating  observations' ID, Age, Gender, Condition in A/B/c testing AND score.}
\usage{
packageData()
}
\value{
A data frame
}
\description{
Package's data calculation
Calculating  observations' ID, Age, Gender, Condition in A/B/c testing AND score.
}
\examples{
Data <- packageData()

}
