% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedT.R
\name{pairedT}
\alias{pairedT}
\title{Paired samples t-test}
\usage{
pairedT(DV, IDV, ID)
}
\arguments{
\item{DV}{A vector with the independent variable}

\item{IDV}{A vector with the dependent variable}

\item{ID}{The ID of the participants}
}
\value{
A list with the descriptive statistics, the model, effect size (if the model is significant and a figure)
}
\description{
The function conducts paired sample t-test. The data set should be with exact n observations per ID, otherwise the figure's and descriptive statistics' results will be inaccurate.
Based on stats::t.test()
}
\examples{
pairedT(theData$Score, theData$Gender, theData$ID)

}
