% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmAnova.R
\name{rmAnova}
\alias{rmAnova}
\title{Repeated measures One Way Anova}
\usage{
rmAnova(DV, IDV, ID, Correction = "BH", Parametric = TRUE)
}
\arguments{
\item{DV}{A vector with the dependent variable}

\item{IDV}{A vector with the independent variable}

\item{ID}{The identity of the observations}

\item{Correction}{The type of correction for post hoc (default is Benjamini, Y., and Hochberg, Y. (1995)) for more details go to the function pairwise.t.test}

\item{Parametric}{If FALSE the test is Friedman test with Wilcoxon Sum Rank tests for post hoc pairwise comparisons}
}
\value{
A list with the following components:

Descriptive_statistics: Descriptive statistics with the Mean, standard deviation, Median and N

Model_summary: Model's summary

Effect_size: eta square effect size (if the model is significant and parametric)

Post_hoc: Pairwise comparisons (if the model is significant)

Figure
}
\description{
This function conducts repeated measures ANOVA. If the model is significant, an effect size and pairwise comparisons also be produced. The data set should be with exact n observations per ID, otherwise the figure's and descriptive statistics' results will be inaccurate
Based on stats::aov()
}
\examples{
rmAnova(theData$Score, theData$Condition, theData$ID)

}
